Moves enumerators
