+++ app/src/main/java/angelosz/catbattlegame/domain/enums/CombatStage.kt
@@ -1,8 +1,9 @@
-package angelosz.catbattlegame.ui.combat
+package angelosz.catbattlegame.domain.enums
 
 enum class CombatStage{
     PLAYER_TURN,
     SELECTING_TARGETS,
     TARGETS_SELECTED,
+    ENDING_TURN,
     AI_TURN
 }
